# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nuclom API",
    "version": "1.0.0",
    "description": "Nuclom is a video collaboration platform that helps teams organize, share, and collaborate on video content. This API provides programmatic access to videos, organizations, series, comments, and more.",
    "contact": {
      "name": "Nuclom Support",
      "url": "https://nuclom.com/support"
    },
    "license": {
      "name": "Elastic License 2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    }
  },
  "servers": [
    {
      "url": "https://nuclom.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/organizations/{id}/audit-logs": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get audit-log",
        "operationId": "getOrganizationAudit-log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create audit-log",
        "operationId": "createOrganizationAudit-log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/organizations/{id}/roles/{roleId}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Delete rol",
        "operationId": "deleteOrganizationRol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "description": "RoleId ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get rol",
        "operationId": "getOrganizationRol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "description": "RoleId ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Patch rol",
        "operationId": "patchOrganizationRol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "description": "RoleId ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication token from login or API key"
      }
    }
  }
}